studytrack-backend/requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
requests==2.32.5
urllib3==2.5.0
Flask
python-dotenv
supabase
requests
